refactor: Extract form settings components
