+++ packages/form-builder/src/settings/index.js
@@ -1,0 +1,11 @@
+import DonationGoalSettings from './donation-goal'
+import FormFieldSettings from './form-fields'
+import FormTitleSettings from './form-title'
+import OfflineDonationsSettings from './offline-donation'
+
+export {
+    DonationGoalSettings,
+    FormFieldSettings,
+    FormTitleSettings,
+    OfflineDonationsSettings,
+}
